feat(fe): implement tabs page with parallel route
