+++ apps/frontend/app/(client)/(main)/course/[courseId]/assignment/[assignmentId]/@tabs/problem/page.tsx
@@ -1,0 +1,3 @@
+export default function AssignmentProblem() {
+  return <div>Problem</div>
+}
