+++ apps/frontend/app/(client)/(main)/course/[courseId]/assignment/[assignmentId]/layout.tsx
@@ -1,10 +1,8 @@
 import { auth } from '@/libs/auth'
-import { fetcher, fetcherWithAuth, getStatusWithStartEnd } from '@/libs/utils'
-import { dateFormatter } from '@/libs/utils'
-import type { Assignment } from '@/types/type'
+import { fetcher, fetcherWithAuth } from '@/libs/utils'
 import { AssignmentTabs } from './_components/AssignmentTabs'
 
-interface ContestDetailProps {
+interface AssignmentDetailProps {
   params: {
     assignmentId: string
     courseId: string
@@ -12,35 +10,22 @@ interface ContestDetailProps {
   tabs: React.ReactNode
 }
 
-export default async function Layout({ params, tabs }: ContestDetailProps) {
+export default async function Layout({ params, tabs }: AssignmentDetailProps) {
   const { assignmentId, courseId } = params
   const session = await auth()
 
-  const res = await (session ? fetcherWithAuth : fetcher).get(
+  const res = await (session ? fetcherWithAuth : fetcher).head(
     `assignment/${assignmentId}`
   )
-  if (res.ok) {
-    const assignment: Assignment = await res.json()
-    const formattedStartTime = dateFormatter(
-      assignment.startTime,
-      'YYYY-MM-DD HH:mm:ss'
-    )
-    const formattedEndTime = dateFormatter(
-      assignment.endTime,
-      'YYYY-MM-DD HH:mm:ss'
-    )
-    const assignmentStatus = getStatusWithStartEnd(
-      formattedStartTime,
-      formattedEndTime
-    )
 
+  if (res.ok) {
     return (
       <article>
         <AssignmentTabs assignmentId={assignmentId} courseId={courseId} />
         {tabs}
-        {assignmentStatus.toString()}
       </article>
     )
   }
+
   return <p className="text-center">No Results.</p>
 }
